bank/realtime/requirements.txt
Faker==37.3.0
numpy==2.3.0
pandas==2.3.0
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
tzdata==2025.2
